requirements.txt
pyaudio
SpeechRecognition
torch
git+https://github.com/openai/whisper.git
discord.py[voice]
python-dotenv